+++ src/sp_validation/__init__.py
@@ -9,6 +9,7 @@ __all__ = [
     'format',
     'galaxy',
     'glass_mock',
+    'masks',
     'plots',
     'rho_tau',
     'statistics',
